Implemented massive code splitting.
