+++ .gitignore
@@ -51,7 +51,6 @@ jspm_packages
 # Optional REPL history
 .node_repl_history
 dist
-**/src/**/*.d.ts
 packages/ste-core/src/**/*.js
 packages/ste-events/src/**/*.js
 packages/ste-signals/src/**/*.js
