+++ packages/ste-events/src/IEvent.ts
@@ -1,0 +1,10 @@
+import { ISubscribable } from "ste-core";
+import { IEventHandler } from "./IEventHandler";
+
+/**
+ * Models an event with a generic sender and generic argument.
+ */
+
+export interface IEvent<TSender, TArgs>
+  extends ISubscribable<IEventHandler<TSender, TArgs>> {
+}
