+++ packages/ste-events/src/index.ts
@@ -1,2 +1,17 @@
-export { EventDispatcher, EventHandlingBase, EventList, NonUniformEventList } from "./events";
-export { IEventHandler, IEventHandling, IEvent } from "./definitions";
+import { EventDispatcher } from "./EventDispatcher";
+import { EventHandlingBase } from "./EventHandlingBase";
+import { EventList } from "./EventList";
+import { IEvent } from "./IEvent";
+import { IEventHandler } from "./IEventHandler";
+import { IEventHandling } from "./IEventHandling";
+import { NonUniformEventList } from "./NonUniformEventList";
+
+export {
+    IEventHandler,
+    IEventHandling,
+    IEvent,
+    EventDispatcher,
+    EventHandlingBase,
+    EventList,
+    NonUniformEventList,
+};
