+++ packages/ste-signals/src/ISignalHandler.ts
@@ -1,0 +1,8 @@
+import { IEventManagement } from "ste-core";
+
+/**
+ * Event handler function without arguments
+ */
+export interface ISignalHandler {
+  (ev: IEventManagement): void;
+}
