+++ packages/ste-signals/src/ISignalHandling.ts
@@ -1,0 +1,7 @@
+import { IBaseEventHandling } from "ste-core";
+import { ISignalHandler } from "./ISignalHandler";
+
+/**
+ * Indicates the object is capable of handling named signals.
+ */
+export interface ISignalHandling extends IBaseEventHandling<ISignalHandler> { }
