+++ packages/ste-signals/src/index.ts
@@ -1,2 +1,11 @@
-export { SignalDispatcher, SignalHandlingBase, SignalList } from "./signals";
-export { ISignalHandling, ISignal, ISignalHandler } from "./definitions";
+import { ISignal } from "./ISignal";
+import { ISignalHandler } from "./ISignalHandler";
+import { ISignalHandling } from "./ISignalHandling";
+import { SignalDispatcher } from "./SignalDispatcher";
+import { SignalHandlingBase } from "./SignalHandlingBase";
+import { SignalList } from "./SignalList";
+
+export { 
+    ISignalHandling, ISignal, ISignalHandler,
+    SignalDispatcher, SignalHandlingBase, SignalList 
+}
