+++ packages/ste-simple-events/src/index.ts
@@ -1,11 +1,17 @@
+import { ISimpleEvent } from "./ISimpleEvent";
+import { SimpleEventDispatcher } from "./SimpleEventDispatcher";
+import { SimpleEventHandlingBase } from "./SimpleEventHandlingBase";
+import { NonUniformSimpleEventList } from "./NonUniformSimpleEventList";
+import { SimpleEventList } from "./SimpleEventList";
+import { ISimpleEventHandler } from "./ISimpleEventHandler";
+import { ISimpleEventHandling } from "./ISimpleEventHandling";
+
 export {
-  SimpleEventDispatcher,
-  SimpleEventHandlingBase,
-  SimpleEventList,
-  NonUniformSimpleEventList
-} from "./simple-events";
-export {
-  ISimpleEventHandling,
-  ISimpleEvent,
-  ISimpleEventHandler
-} from "./definitions";
+    SimpleEventDispatcher,
+    SimpleEventHandlingBase,
+    SimpleEventList,
+    NonUniformSimpleEventList,
+    ISimpleEvent,
+    ISimpleEventHandler,
+    ISimpleEventHandling,
+};
